Fixed showsMyLocationButton
