+++ components/Signup/SignupSection.js
@@ -1,34 +1,17 @@
 import React, {Component} from 'react';
-import Dimensions from 'Dimensions';
 import {StyleSheet, 
         View, 
         Text
 } from 'react-native';
+import SignupSectionStyles from './SignupSectionStyles'
 
 export default class SignupSection extends Component {
     render() {
       return (
-        <View style={styles.container}>
-          <Text style={styles.text}>Create Account</Text>
-          <Text style={styles.text}>Forgot Password?</Text>
+        <View style={SignupSectionStyles.container}>
+          <Text style={SignupSectionStyles.text}>Create Account</Text>
+          <Text style={SignupSectionStyles.text}>Forgot Password?</Text>
         </View>
       );
     }
 }
-
-const DEVICE_WIDTH = Dimensions.get('window').width;
-const DEVICE_HEIGHT = Dimensions.get('window').height;
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    top: 65,
-    width: DEVICE_WIDTH,
-    flexDirection: 'row',
-    justifyContent: 'space-around',
-  },
-  text: {
-    color: 'white',
-    backgroundColor: 'transparent',
-  },
-});
