+++ components/Signup/SignupSectionStyles.js
@@ -1,0 +1,19 @@
+import { StyleSheet } from "react-native"
+import Dimensions from 'Dimensions';
+
+const DEVICE_WIDTH = Dimensions.get('window').width;
+const DEVICE_HEIGHT = Dimensions.get('window').height;
+
+export default StyleSheet.create({
+  container: {
+    flex: 1,
+    top: 65,
+    width: DEVICE_WIDTH,
+    flexDirection: 'row',
+    justifyContent: 'space-around',
+  },
+  text: {
+    color: 'white',
+    backgroundColor: 'transparent',
+  },
+});
